requirements.txt
Django==1.11.13
django-filter==1.1.0
pytz==2018.4
